Rename index.ts to lit-html.ts
