+++ src/directives/async-append.ts
@@ -12,7 +12,7 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {directive, Directive, NodePart} from '../index.js';
+import {directive, Directive, NodePart} from '../lit-html.js';
 
 /**
  * A directive that renders the items of an async iterable[1], appending new
