+++ src/test/directives/async-replace_test.ts
@@ -13,8 +13,8 @@
  */
 
 import {asyncReplace} from '../../directives/async-replace.js';
-import {html} from '../../index.js';
 import {render} from '../../lib/render.js';
+import {html} from '../../lit-html.js';
 import {TestAsyncIterable} from '../lib/test-async-iterable.js';
 import {stripExpressionMarkers} from '../test-utils/strip-markers.js';
