+++ src/test/directives/unsafe-html_test.ts
@@ -13,8 +13,8 @@
  */
 
 import {unsafeHTML} from '../../directives/unsafe-html.js';
-import {html} from '../../index.js';
 import {render} from '../../lib/render.js';
+import {html} from '../../lit-html.js';
 import {stripExpressionMarkers} from '../test-utils/strip-markers.js';
 
 const assert = chai.assert;
