+++ src/test/lib/modify-template_test.ts
@@ -12,9 +12,9 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {html, templateFactory} from '../../index.js';
 import {insertNodeIntoTemplate, removeNodesFromTemplate} from '../../lib/modify-template.js';
 import {render} from '../../lib/render.js';
+import {html, templateFactory} from '../../lit-html.js';
 import {stripExpressionMarkers} from '../test-utils/strip-markers.js';
 
 const assert = chai.assert;
